[UPSTREAM?] Fix out-of-bounds read in QmlJsLexer
This only happens when the string is not NUL-terminated. I found this
by running some of the QtQml tests compiled for CHERI-RISC-V.
CHERI-enabled architectures (e.g. Arm Morello) provide fine-grained
bounds and permissions for all pointers, so this out-of-bounds read
can be detected. The problem happens when evaluating
eng.evaluate(QStringLiteral("...."));
since the trailing NUL character will not be included in the QString
bounds when created from a QStringLiteral().

Change-Id: I6ac2a6c8a7aa6fb7db8dd8659ddfac75dea2d080