Make the host CompiledData::String compatible with CHERI QStringData
This file is also used on for the host in order to bytecode-compile
.qml files ahead of time. The string data is reinterpreted as a
QStringData, so the alignment and layout needs to be compatible with
all supported architectures (including CHERI). While the offset will
never be a valid capability (since it's a static string) we do need to
ensure alignment since we will use a capability load to load it.

This commit ensures that alignment and size of the String struct is
compatible with CHERI even when compiled e.g. on a x86 host. Fixes
some of the qml tests.

Change-Id: I41ae4f7dde8964244c821c9f259edb4bdfe6df2d