qintrusivelvist_p.h: Silence nullpointer subtraction warning
Fixing this class to avoid the UB here is not worth the effort. Using a
pointer which is not nullptr "works", but does not change the underlying
fact that only the offsetof macro is allowed this trickery. So instead
of hiding the issue, we explicitly silence it for now.

Change-Id: I55f456049e78f24434e3494fc94c83681e5f46c9
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>
(cherry picked from commit 83ed0e4cd90e384e6d735066e19c7f505189f98d)